회원 탈퇴시, 비밀번호 재입력 기능 구현 -main.js 삭제, unsub.js 생성 -mypageController 수정 및 unsubscribe.jsp 생성

+++ BookieProject/src/main/java/com/project/bookie/controller/MyPageController.java
@@ -2,9 +2,11 @@ package com.project.bookie.controller;
 
 import java.util.List;
 
+import org.apache.ibatis.annotations.Param;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,6 +24,8 @@ import com.project.bookie.service.UserService;
 public class MyPageController {
 	@Autowired
 	UserService userService;
+	@Autowired
+	BCryptPasswordEncoder passwordEncoder;
 
 	@GetMapping("/info")
 	public String showMypage(Model m) {
@@ -51,14 +55,28 @@ public class MyPageController {
 		return "/myPage/myPage";
 	}
 
-	@PostMapping(value = "/delete", produces = "text/plain; charset=utf8")
+	@GetMapping("/unsubscribe")
+	public String unsubscribeChk(Model m) {
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		String uEmail = auth.getName();
+		m.addAttribute("uEmail", uEmail);
+		return "myPage/unsubscribe";
+	}
+
+	@PostMapping(value = "/unsubscribe", produces = "text/plain; charset=utf-8")
 	@ResponseBody
-	public String deleteUser(@RequestParam("uEmail") String uEmail) {
-		try {
+	public String deleteUser(@RequestParam(value = "uEmail") String uEmail, @RequestParam(value = "pwd") String pwd) {
+		System.out.println(uEmail);
+		System.out.println(pwd);
+		String db_pwd = userService.getUserPwd(uEmail); // db의 user pwd
+		if (passwordEncoder.matches(pwd, db_pwd) == true) {
+			System.out.println("success!");
 			userService.deleteUserinSystem(userService.getUserIdByEmail(uEmail));
-		} catch (Exception e) {
+			System.out.println("return");
+			return "true";
+		} else {
+			System.out.println("false");
 			return "false";
 		}
-		return "true";
 	}
 }
